srcs/requirement.txt
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
Django==2.1.7
django-slack==5.13.0
djangorestframework==3.9.2
djongo==1.2.32
dojango==0.5.8
future==0.17.1
idna==2.8
oauthlib==3.0.1
pymongo==3.7.2
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
slackclient==1.3.1
sqlparse==0.2.4
urllib3==1.24.1
websocket-client==0.54.0
